requirements.txt
Django
djangorestframework
names
requests
coreapi
